requirements.txt
# Face Recognition System Requirements
# Python 3.8+
tensorflow>=2.10.0
opencv-python>=4.7.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyserial>=3.5
